iOS Exchange Redesign. Other Minor Updates

+++ ios/batm/Entities/CoinSettings.swift
@@ -5,7 +5,7 @@ struct CoinSettings: Equatable {
   var type: CustomCoinType
   var txFee: Decimal
   var recallFee: Decimal?
-  var byteFee: Decimal?
+  var byteFee: Int?
   var gasPrice: Int?
   var gasLimit: Int?
   var profitExchange: Decimal
